Rename the kernel to linux-arch4ec2.
The intent here is to have a different name so that people need not be
manually paying attention to avoiding accidentally upgrading to a
different kernel. At some point I should look into getting this into
AUR and such, but for now I just want something self-contained that
works simply for people.