Use proper repository extension and stop manual symlinking.
A newer version of repo-add was validating the extension. It was
previously very unclear to me what the naming convention was supposed
to be, and lots of confusion ensued, which lead to the creation of the
file and manual symlinking. The new repo-add just does the right thing
if given *.db.tar.gz.